Adds new ways to create event sync.
In order to add more functionality to the project, this commit adds 2
new ways to create EventSync, and a way to reset the starting time.

EventSync can now be created with a given amount of time having already
passed, both through ticks and a Duration. Achieved through the
`from_starting_time` and `from_starting_tick` methods.