Changes EventSync to store and Arc RwLock of its data
In order to consistently connect instances of EventSync, this commit
moves all of the EventSync data to an InnerEventSync type. EventSync now
contains a field that wraps InnerEventSync in an Arc RwLock. This will
allow for consistent communication between EventSyncs when any internal
data is changed between them. Such as being paused, resetting the
starting time, or changing the tickrate.

In addition, this commit makes some documentation changes, and adds an
extra argument to creating EventSync from ticks and time to start it
paused.